+++ src/main/java/de/mari_023/ae2wtlib/terminal/Icon.java
@@ -49,8 +49,6 @@ public record Icon(int x, int y, int width, int height, Texture texture) {
     public static final Icon UPGRADE_BACKGROUND_SCROLLING_MIDDLE = new Icon(48, 85, 29, 18);
     public static final Icon UPGRADE_BACKGROUND_SCROLLING_BOTTOM = new Icon(48, 103, 29, 25);
 
-    public static final Icon SINGULARITY_BACKGROUND = new Icon(100, 105, 28, 23);
-
     private Icon(int x, int y) {
         this(x, y, 16, 16);
     }