+++ src/main/java/de/mari_023/ae2wtlib/wat/WATScreen.java
@@ -6,9 +6,7 @@ import net.minecraft.world.entity.player.Inventory;
 import appeng.client.gui.me.patternaccess.PatternAccessTermScreen;
 import appeng.client.gui.style.ScreenStyle;
 import appeng.client.gui.widgets.ToolboxPanel;
-import appeng.menu.SlotSemantics;
 
-import de.mari_023.ae2wtlib.terminal.ScrollingUpgradesPanel;
 import de.mari_023.ae2wtlib.terminal.WTMenuHost;
 import de.mari_023.ae2wtlib.wut.IUniversalTerminalCapable;
 
@@ -18,11 +16,9 @@ public class WATScreen extends PatternAccessTermScreen<WATMenu> implements IUniv
         if (getMenu().isWUT())
             addToLeftToolbar(cycleTerminalButton());
 
-        widgets.add("scrollingUpgrades",
-                new ScrollingUpgradesPanel(menu.getSlots(SlotSemantics.UPGRADE), menu.getHost(), widgets));
+        addUpgradePanel(widgets, getMenu());
         if (getMenu().getToolbox().isPresent())
             widgets.add("toolbox", new ToolboxPanel(style, getMenu().getToolbox().getName()));
-        addSingularityPanel(widgets, getMenu());
     }
 
     @Override