+++ src/main/java/de/mari_023/ae2wtlib/wct/WCTScreen.java
@@ -10,7 +10,6 @@ import net.minecraft.world.inventory.Slot;
 
 import appeng.client.gui.me.items.CraftingTermScreen;
 import appeng.client.gui.style.ScreenStyle;
-import appeng.menu.SlotSemantics;
 
 import de.mari_023.ae2wtlib.TextConstants;
 import de.mari_023.ae2wtlib.terminal.*;
@@ -38,10 +37,7 @@ public class WCTScreen extends CraftingTermScreen<WCTMenu> implements IUniversal
         trashButton.setMessage(TextConstants.TRASH);
 
         widgets.add("player", new PlayerEntityWidget(Objects.requireNonNull(Minecraft.getInstance().player)));
-        addSingularityPanel(widgets, getMenu());
-
-        widgets.add("scrollingUpgrades",
-                new ScrollingUpgradesPanel(menu.getSlots(SlotSemantics.UPGRADE), menu.getHost(), widgets));
+        addUpgradePanel(widgets, getMenu());
     }
 
     private void setMagnetMode() {