+++ src/main/java/de/mari_023/ae2wtlib/wet/WETScreen.java
@@ -5,9 +5,7 @@ import net.minecraft.world.entity.player.Inventory;
 
 import appeng.client.gui.me.items.PatternEncodingTermScreen;
 import appeng.client.gui.style.ScreenStyle;
-import appeng.menu.SlotSemantics;
 
-import de.mari_023.ae2wtlib.terminal.ScrollingUpgradesPanel;
 import de.mari_023.ae2wtlib.terminal.WTMenuHost;
 import de.mari_023.ae2wtlib.wut.IUniversalTerminalCapable;
 
@@ -16,9 +14,7 @@ public class WETScreen extends PatternEncodingTermScreen<WETMenu> implements IUn
         super(container, playerInventory, title, style);
         if (getMenu().isWUT())
             addToLeftToolbar(cycleTerminalButton());
-        addSingularityPanel(widgets, getMenu());
-        widgets.add("scrollingUpgrades",
-                new ScrollingUpgradesPanel(menu.getSlots(SlotSemantics.UPGRADE), menu.getHost(), widgets));
+        addUpgradePanel(widgets, getMenu());
     }
 
     @Override