+++ src/main/java/de/mari_023/ae2wtlib/wut/IUniversalTerminalCapable.java
@@ -1,5 +1,6 @@
 package de.mari_023.ae2wtlib.wut;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.jetbrains.annotations.Contract;
@@ -10,13 +11,13 @@ import appeng.client.Hotkeys;
 import appeng.client.gui.WidgetContainer;
 import appeng.core.network.serverbound.HotkeyPacket;
 import appeng.menu.AEBaseMenu;
-import appeng.menu.slot.AppEngSlot;
+import appeng.menu.SlotSemantics;
 
 import de.mari_023.ae2wtlib.AE2wtlibSlotSemantics;
 import de.mari_023.ae2wtlib.TextConstants;
 import de.mari_023.ae2wtlib.networking.CycleTerminalPacket;
 import de.mari_023.ae2wtlib.terminal.IconButton;
-import de.mari_023.ae2wtlib.terminal.SingularityPanel;
+import de.mari_023.ae2wtlib.terminal.ScrollingUpgradesPanel;
 import de.mari_023.ae2wtlib.terminal.WTMenuHost;
 
 public interface IUniversalTerminalCapable {
@@ -59,14 +60,15 @@ public interface IUniversalTerminalCapable {
     }
 
     /**
-     * creates and adds the background for the singularity
+     * creates and adds the upgrade panel
      * 
      * @param widgets the WidgetContainer where the widget will be added
      * @param menu    the menu corresponding to this screen
      */
-    default void addSingularityPanel(WidgetContainer widgets, AEBaseMenu menu) {
-        widgets.add("singularity",
-                new SingularityPanel((AppEngSlot) menu.getSlots(AE2wtlibSlotSemantics.SINGULARITY).getFirst(),
-                        () -> getHost().getUpgrades()));
+    default void addUpgradePanel(WidgetContainer widgets, AEBaseMenu menu) {
+        var upgrades = new ArrayList<>(menu.getSlots(AE2wtlibSlotSemantics.SINGULARITY));
+        upgrades.addAll(menu.getSlots(SlotSemantics.UPGRADE));
+        widgets.add("scrollingUpgrades",
+                new ScrollingUpgradesPanel(upgrades, getHost(), widgets, () -> getHost().getUpgrades()));
     }
 }